Added mongodb for product.

+++ src/main/java/com/departmental/store/api/product/controller/request/ProductCreateRequest.java
@@ -3,14 +3,14 @@ package com.departmental.store.api.product.controller.request;
 public class ProductCreateRequest {
 
     private String name;
-    private long price;
+    private float price;
     private int quantity;
 
     public String getName() {
         return name;
     }
 
-    public long getPrice() {
+    public float getPrice() {
         return price;
     }
